+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0+10 - Working dynamic boundaries on web platform
+
+ * The web platform can now compute the Widget's width and height on it's own. No more need for `SizedBox` wrap.
+ * Added dynamic generation of the required JavaScript code. Please update your `web/index.html` according to [README.md] and run `flutter pub run katex_flutter:main`
+
 ## 2.0.2+8 - Fixed equations displayed twice
 
  * Fixed some equations rendered twice due to inpropper MathML compatibility